Fix x64 build error

+++ PowerEditor/src/Notepad_plus.cpp
@@ -978,7 +978,7 @@ int Notepad_plus::getHtmlXmlEncoding(const TCHAR *fileName) const
 	_invisibleEditView.execute(SCI_APPENDTEXT, lenFile, reinterpret_cast<LPARAM>(data));
 
 	const char *encodingAliasRegExpr = "[a-zA-Z0-9_-]+";
-	const int encodingStrLen = 128;
+	const size_t encodingStrLen = 128;
 	if (langT == L_XML)
 	{
 		// find encoding by RegExpr
@@ -1005,7 +1005,7 @@ int Notepad_plus::getHtmlXmlEncoding(const TCHAR *fileName) const
             startPos = int(_invisibleEditView.execute(SCI_GETTARGETSTART));
 			endPos = _invisibleEditView.execute(SCI_GETTARGETEND);
 			
-			int len = endPos - startPos;
+			size_t len = endPos - startPos;
 			if (len >= encodingStrLen)
 			{
 				return -1;
@@ -1049,7 +1049,7 @@ int Notepad_plus::getHtmlXmlEncoding(const TCHAR *fileName) const
         startPos = int(_invisibleEditView.execute(SCI_GETTARGETSTART));
 		endPos = _invisibleEditView.execute(SCI_GETTARGETEND);
 
-		int len = endPos - startPos;
+		size_t len = endPos - startPos;
 		if (len >= encodingStrLen)
 		{
 			return -1;
